requirements.txt
pandas==2.0.3
Pillow==9.5.0
Pillow==10.0.0
scikit_learn==1.3.0
streamlit==1.24.1
tabulate==0.9.0
